feat(quaint): split sqlite connector into native and wasm submodules
