+++ quaint/Cargo.toml
@@ -70,8 +70,8 @@ mysql-connector = ["mysql", "mysql_async", "tokio/time", "lru-cache"]
 mysql = ["chrono/std"]
 
 pooled = ["mobc"]
-sqlite-connector = ["sqlite", "rusqlite", "tokio/sync"]
-sqlite = []
+sqlite-connector = ["sqlite", "rusqlite/bundled", "tokio/sync"]
+sqlite = ["rusqlite"]
 
 fmt-sql = ["sqlformat"]
 
@@ -127,7 +127,7 @@ branch = "vendored-openssl"
 
 [dependencies.rusqlite]
 version = "0.29"
-features = ["chrono", "bundled", "column_decltype"]
+features = ["chrono", "column_decltype"]
 optional = true
 
 [target.'cfg(not(any(target_os = "macos", target_os = "ios")))'.dependencies.tiberius]
