+++ quaint/src/connector/sqlite/wasm/common.rs
