+++ quaint/src/connector/sqlite/wasm/mod.rs
@@ -1,0 +1,4 @@
+///! Wasm-compatible definitions for the SQLite connector.
+/// /// This module is only available with the `sqlite` feature.
+pub(crate) mod common;
+pub mod error;
