+++ quaint/src/error.rs
@@ -8,7 +8,7 @@ use std::time::Duration;
 
 pub use crate::connector::mysql::MysqlError;
 pub use crate::connector::postgres::PostgresError;
-// pub use crate::connector::sqlite::SqliteError;
+pub use crate::connector::sqlite::SqliteError;
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum DatabaseConstraint {
